+++ frontend/src/embedding-sdk-bundle/components/public/StaticQuestion/StaticQuestion.unit.spec.tsx
@@ -1,11 +1,6 @@
-import userEvent from "@testing-library/user-event";
-
 import { setupEnterpriseOnlyPlugin } from "__support__/enterprise";
 import {
-  findRequests,
-  setupAdhocQueryMetadataEndpoint,
   setupAlertsEndpoints,
-  setupCardDataset,
   setupCardEndpoints,
   setupCardQueryEndpoints,
   setupCardQueryMetadataEndpoint,
@@ -26,65 +21,31 @@ import {
   waitForLoaderToBeRemoved,
   within,
 } from "__support__/ui";
+import { addAlertModalTests } from "embedding-sdk-bundle/components/public/question/shared-tests/alert-modal.spec";
+import { addAlertsButtonTests } from "embedding-sdk-bundle/components/public/question/shared-tests/alerts-button.spec";
+import type { SetupOpts } from "embedding-sdk-bundle/components/public/question/shared-tests/constants.spec";
+import {
+  TEST_COLUMN,
+  TEST_DATASET,
+  TEST_DB,
+  TEST_TABLE,
+} from "embedding-sdk-bundle/components/public/question/shared-tests/constants.spec";
+import { addQueryPropTests } from "embedding-sdk-bundle/components/public/question/shared-tests/query-prop.spec";
 import { renderWithSDKProviders } from "embedding-sdk-bundle/test/__support__/ui";
 import { createMockSdkConfig } from "embedding-sdk-bundle/test/mocks/config";
 import { setupSdkState } from "embedding-sdk-bundle/test/server-mocks/sdk-init";
-import { reinitialize } from "metabase/plugins";
-import { utf8_to_b64url } from "metabase/utils/encoding";
-import type { CardId, CollectionType, TokenFeatures } from "metabase-types/api";
 import {
   createMockCard,
   createMockCardQueryMetadata,
   createMockCollection,
-  createMockColumn,
-  createMockDatabase,
-  createMockDataset,
-  createMockDatasetData,
-  createMockTable,
   createMockTokenFeatures,
   createMockUser,
 } from "metabase-types/api/mocks";
-import { createMockNotification } from "metabase-types/api/mocks/notification";
 
 import { StaticQuestion, StaticQuestionInternal } from "./StaticQuestion";
 
-const TEST_DB_ID = 1;
-const TEST_DB = createMockDatabase({ id: TEST_DB_ID });
-
-const TEST_TABLE_ID = 1;
-const TEST_TABLE = createMockTable({ id: TEST_TABLE_ID, db_id: TEST_DB_ID });
-
-const TEST_COLUMN = createMockColumn({
-  display_name: "Test Column",
-  name: "Test Column",
-});
-
-const TEST_DATASET = createMockDataset({
-  data: createMockDatasetData({
-    cols: [TEST_COLUMN],
-    rows: [["Test Row"]],
-  }),
-});
-
-const TEST_CARD_ID: CardId = 1 as const;
-
+const TEST_CARD_ID = 1;
 const USER_ID = 999;
-
-interface SetupOpts {
-  title?: string | boolean;
-  withChartTypeSelector?: boolean;
-  withDownloads?: boolean;
-  withAlerts?: boolean;
-  isEmailSetup?: boolean;
-  canManageSubscriptions?: boolean;
-  isSuperuser?: boolean;
-  isModel?: boolean;
-  notifications?: ReturnType<typeof createMockNotification>[];
-  collectionType?: CollectionType;
-  tokenFeatures?: Partial<TokenFeatures>;
-  enterprisePlugins?: Parameters<typeof setupEnterpriseOnlyPlugin>[0][];
-  children?: React.ReactNode;
-}
 
 const setup = async ({
   title,
@@ -103,7 +64,7 @@ const setup = async ({
 }: SetupOpts = {}) => {
   setupNotificationChannelsEndpoints({
     email: { configured: isEmailSetup },
-  } as any);
+  });
 
   const user = createMockUser({
     id: USER_ID,
@@ -192,7 +153,12 @@ const setup = async ({
   await waitForLoaderToBeRemoved();
 };
 
+addQueryPropTests({ Component: StaticQuestionInternal });
+
 describe("StaticQuestion", () => {
+  addAlertsButtonTests(setup, { customComponent: StaticQuestion.AlertsButton });
+  addAlertModalTests(setup, { userId: USER_ID });
+
   beforeAll(() => {
     mockGetBoundingClientRect();
   });
@@ -272,182 +238,6 @@ describe("StaticQuestion", () => {
       ).toBeVisible();
     });
 
-    describe("alerts button with different Metabase version configurations", () => {
-      beforeEach(() => {
-        reinitialize();
-      });
-
-      // eslint-disable-next-line jest/no-disabled-tests -- Fix this in EMB-1184, when we can test SDK with API keys
-      describe.skip("OSS (Open Source Software)", () => {
-        it("should not show the alert button in OSS regardless of settings", async () => {
-          // Don't setup enterprise plugin for OSS
-          await setup({
-            withAlerts: true,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            isModel: false,
-            tokenFeatures: {}, // No embedding_sdk feature
-          });
-
-          expect(
-            within(screen.getByRole("gridcell")).getByText("Test Row"),
-          ).toBeVisible();
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-      });
-
-      // eslint-disable-next-line jest/no-disabled-tests -- Fix this in EMB-1184, when we can test SDK with API keys
-      describe.skip("EE (Enterprise Edition) without embedding_sdk token feature", () => {
-        it("should not show the alert button when plugin is enabled but token feature is missing", async () => {
-          await setup({
-            withAlerts: true,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            isModel: false,
-            enterprisePlugins: ["sdk_notifications"],
-            tokenFeatures: {}, // No embedding_sdk feature
-          });
-
-          expect(
-            within(screen.getByRole("gridcell")).getByText("Test Row"),
-          ).toBeVisible();
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-      });
-
-      describe("EE with embedding_sdk token feature", () => {
-        it("should show the alert button when plugin and token feature are both enabled", async () => {
-          await setup({
-            withAlerts: true,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            isModel: false,
-            enterprisePlugins: ["sdk_notifications"],
-          });
-
-          expect(
-            within(screen.getByRole("gridcell")).getByText("Test Row"),
-          ).toBeVisible();
-          expect(
-            await screen.findByRole("button", { name: "Alerts" }),
-          ).toBeVisible();
-        });
-
-        it("should show the alert button for custom layouts when withAlerts is true", async () => {
-          await setup({
-            withAlerts: true,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            isModel: false,
-            enterprisePlugins: ["sdk_notifications"],
-            children: (
-              <div>
-                <span>Custom Layout</span>
-                <StaticQuestion.AlertsButton />
-              </div>
-            ),
-          });
-
-          expect(screen.getByText("Custom Layout")).toBeVisible();
-          expect(screen.getByRole("button", { name: "Alerts" })).toBeVisible();
-        });
-
-        it("should not show the alert button for custom layouts when withAlerts is false", async () => {
-          await setup({
-            withAlerts: false,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            isModel: false,
-            enterprisePlugins: ["sdk_notifications"],
-            children: (
-              <div>
-                <span>Custom Layout</span>
-                <StaticQuestion.AlertsButton />
-              </div>
-            ),
-          });
-
-          expect(screen.getByText("Custom Layout")).toBeVisible();
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-
-        it("should not show the alert button when withAlerts is false", async () => {
-          await setup({
-            withAlerts: false,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            enterprisePlugins: ["sdk_notifications"],
-          });
-
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-
-        it("should not show the alert button when email is not configured", async () => {
-          await setup({
-            withAlerts: true,
-            isEmailSetup: false,
-            canManageSubscriptions: true,
-            enterprisePlugins: ["sdk_notifications"],
-          });
-
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-
-        it("should not show the alert button when user cannot manage subscriptions and is not admin", async () => {
-          await setup({
-            withAlerts: true,
-            isEmailSetup: true,
-            isSuperuser: false,
-            canManageSubscriptions: false,
-            enterprisePlugins: ["sdk_notifications", "application_permissions"],
-            tokenFeatures: { embedding_sdk: true, advanced_permissions: true },
-          });
-
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-
-        it("should not show the alert button for models", async () => {
-          await setup({
-            withAlerts: true,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            isModel: true,
-            enterprisePlugins: ["sdk_notifications"],
-          });
-
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-
-        it("should not show the alert button for analytics collection", async () => {
-          await setup({
-            withAlerts: true,
-            isEmailSetup: true,
-            canManageSubscriptions: true,
-            collectionType: "instance-analytics",
-            enterprisePlugins: ["sdk_notifications"],
-          });
-
-          expect(
-            screen.queryByRole("button", { name: "Alerts" }),
-          ).not.toBeInTheDocument();
-        });
-      });
-    });
-
     it("should show the TopBar when multiple UI elements are provided", async () => {
       await setup({
         title: "My Title",
@@ -463,160 +253,5 @@ describe("StaticQuestion", () => {
       expect(screen.getByText("My Title")).toBeVisible();
       expect(screen.getByTestId("chart-type-selector-button")).toBeVisible();
     });
-  });
-
-  describe("alert modal", () => {
-    it("should show alert list modal when there are existing alerts", async () => {
-      await setup({
-        withAlerts: true,
-        isEmailSetup: true,
-        canManageSubscriptions: true,
-        isModel: false,
-        enterprisePlugins: ["sdk_notifications"],
-        notifications: [createMockNotification()],
-      });
-
-      expect(
-        within(screen.getByRole("gridcell")).getByText("Test Row"),
-      ).toBeVisible();
-
-      await userEvent.click(
-        await screen.findByRole("button", { name: "Alerts" }),
-      );
-
-      // Verify the alert list modal appears, not the create modal
-      const withinModal = within(await findModal());
-
-      // Show the alert list modal title
-      expect(
-        await withinModal.findByRole("heading", { name: "Edit alerts" }),
-      ).toBeVisible();
-
-      // Should show the button to create a new alert in the list modal
-      expect(withinModal.getByText("New alert")).toBeVisible();
-
-      // Should NOT show the create alert form fields
-      expect(
-        withinModal.queryByText("What do you want to be alerted about?"),
-      ).not.toBeInTheDocument();
-    });
-
-    it("should not show email selector on the SDK and use current logged in user as the recipient", async () => {
-      await setup({
-        withAlerts: true,
-        isEmailSetup: true,
-        canManageSubscriptions: true,
-        isModel: false,
-        enterprisePlugins: ["sdk_notifications"],
-      });
-
-      expect(
-        within(screen.getByRole("gridcell")).getByText("Test Row"),
-      ).toBeVisible();
-      await userEvent.click(
-        await screen.findByRole("button", { name: "Alerts" }),
-      );
-
-      const withinModal = within(await findModal());
-      expect(
-        withinModal.getByRole("heading", { name: "New alert" }),
-      ).toBeVisible();
-      expect(
-        withinModal.getByText("What do you want to be alerted about?"),
-      ).toBeVisible();
-      expect(
-        withinModal.getByText("When do you want to check this?"),
-      ).toBeVisible();
-      // Email selector is within this section, checking only the header is fine
-      expect(
-        withinModal.queryByText("Where do you want to send the results?"),
-      ).not.toBeInTheDocument();
-      expect(withinModal.getByText("More options")).toBeVisible();
-      await userEvent.click(withinModal.getByRole("button", { name: "Done" }));
-
-      const createNotificationRequest = (await findRequests("POST")).find(
-        (postRequest) =>
-          postRequest.url === "http://localhost/api/notification",
-      );
-      // Checks that we're using the current logged in user as the sole recipient
-      expect(createNotificationRequest?.body).toMatchObject({
-        handlers: [
-          {
-            channel_type: "channel/email",
-            recipients: [
-              {
-                details: null,
-                type: "notification-recipient/user",
-                user_id: USER_ID,
-              },
-            ],
-          },
-        ],
-      });
-      // So that when we assert this value, we know we won't accidentally match the default mock ID
-      expect(USER_ID).not.toBe(1);
-    });
-  });
-});
-
-async function findModal() {
-  return await screen.findByRole("dialog");
-}
-
-describe("StaticQuestion — query prop", () => {
-  const QUERY_PROP = utf8_to_b64url(
-    JSON.stringify({
-      dataset_query: {
-        database: TEST_DB_ID,
-        type: "query",
-        query: { "source-table": TEST_TABLE_ID },
-      },
-    }),
-  );
-
-  async function setup() {
-    const { state } = setupSdkState();
-
-    setupNotificationChannelsEndpoints({ email: { configured: false } } as any);
-    setupAdhocQueryMetadataEndpoint(
-      createMockCardQueryMetadata({ databases: [TEST_DB] }),
-    );
-    // Needs >1 row so Question.maybeResetDisplay doesn't force scalar.
-    setupCardDataset({
-      dataset: createMockDataset({
-        data: createMockDatasetData({
-          cols: [TEST_COLUMN],
-          rows: [["Test Row"], ["Test Row 2"]],
-        }),
-      }),
-    });
-    setupCollectionByIdEndpoint({
-      collections: [createMockCollection({ id: 1 })],
-    });
-
-    renderWithSDKProviders(<StaticQuestionInternal query={QUERY_PROP} />, {
-      componentProviderProps: { authConfig: createMockSdkConfig() },
-      storeInitialState: state,
-    });
-
-    await waitForLoaderToBeRemoved();
-  }
-
-  beforeAll(() => {
-    mockGetBoundingClientRect();
-  });
-
-  it("should render a visualization when given a valid query base64 string", async () => {
-    await setup();
-
-    expect(screen.getByTestId("query-visualization-root")).toBeVisible();
-    expect(
-      within(screen.getByTestId("table-root")).getByText(
-        TEST_COLUMN.display_name,
-      ),
-    ).toBeVisible();
-    expect(
-      within(screen.getAllByRole("gridcell")[0]).getByText("Test Row"),
-    ).toBeVisible();
   });
 });
